+++ Watchman.Discord/Areas/Protection/Controllers/AntiSpamController.cs
@@ -10,7 +10,7 @@ using Devscord.DiscordFramework.Middlewares.Contexts;
 using Serilog;
 using Watchman.Discord.Areas.Protection.Services;
 using Watchman.Discord.Areas.Protection.Strategies;
-using Watchman.DomainModel.Settings.Services;
+using Watchman.DomainModel.Configuration.Services;
 
 namespace Watchman.Discord.Areas.Protection.Controllers
 {
