+++ Watchman.Discord/Areas/Protection/Strategies/CheckUserSafetyService.cs
@@ -8,8 +8,8 @@ using Watchman.Cqrs;
 using Watchman.Discord.Areas.Protection.Models;
 using Watchman.DomainModel.DiscordServer.Queries;
 using Watchman.DomainModel.Messages.Queries;
-using Watchman.DomainModel.Settings.ConfigurationItems;
-using Watchman.DomainModel.Settings.Services;
+using Watchman.DomainModel.Configuration.ConfigurationItems;
+using Watchman.DomainModel.Configuration.Services;
 using Message = Watchman.DomainModel.Messages.Message;
 
 namespace Watchman.Discord.Areas.Protection.Strategies
