+++ Watchman.DomainModel/Configuration/ConfigurationItem.cs
@@ -1,6 +1,6 @@
 ﻿using Watchman.Integrations.MongoDB;
 
-namespace Watchman.DomainModel.Settings
+namespace Watchman.DomainModel.Configuration
 {
     public class ConfigurationItem : Entity, IAggregateRoot
     {
