+++ Watchman.DomainModel/Configuration/ConfigurationItems/HowManyMessagesInShortTimeToBeSpam.cs
@@ -1,4 +1,4 @@
-﻿namespace Watchman.DomainModel.Settings.ConfigurationItems
+﻿namespace Watchman.DomainModel.Configuration.ConfigurationItems
 {
     public class HowManyMessagesInShortTimeToBeSpam : MappedConfiguration<int>
     {
