+++ Watchman.DomainModel/Configuration/ConfigurationItems/MinAverageMessagesPerWeek.cs
@@ -1,4 +1,4 @@
-﻿namespace Watchman.DomainModel.Settings.ConfigurationItems
+﻿namespace Watchman.DomainModel.Configuration.ConfigurationItems
 {
     public class MinAverageMessagesPerWeek : MappedConfiguration<int>
     {
