+++ Watchman.DomainModel/Configuration/IMappedConfiguration.cs
@@ -1,4 +1,4 @@
-﻿namespace Watchman.DomainModel.Settings
+﻿namespace Watchman.DomainModel.Configuration
 {
     public interface IMappedConfiguration
     {
